C.Problem Solution

+++ Assuit icpc/sheet 1/C.java
@@ -2,8 +2,8 @@ import java.util.Scanner;
 public class C {
     public static void main(String[] args){
         Scanner sc = new Scanner(System.in);
-        long X = sc.nextInt();
-        long Y = sc.nextInt();
+        long X = sc.nextLong();
+        long Y = sc.nextLong();
         System.out.println(X + " + " + Y + " = " + (X + Y));
         System.out.println(X + " * " + Y + " = " + (X * Y));
         System.out.println(X + " - " + Y + " = " + (X - Y));
